requirements.txt
PyYAML>=6.0
dbus-python>=1.2.18
